tweak gas/elec settings

+++ src/WebBlog/Data/Metric.cs
@@ -5,7 +5,7 @@ namespace WebBlog.Data
     public class Metric
     {
         public long Id { get; set; }
-        public int? Value { get; set; }
+        public decimal? Value { get; set; }
         public DateTime? Date { get; set; }
         public int? Type { get; set; }
         public string PartitionKey { get; set; }
